src/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# NVIDIA Triton client
tritonclient[all]==2.40.0

# Computer vision and image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# YOLO model support
ultralytics==8.0.196

# Async and HTTP clients
httpx==0.25.2
aiofiles==23.2.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Configuration and utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0 